backend/requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
scikit-learn
xgboost
lightgbm
catboost
tensorflow-cpu
torch
torchvision
pywavelets
neurokit2
wfdb
emd-signal
opencv-python
pillow
tqdm
fastapi
uvicorn
python-multipart
jupyterlab
pdf2image
pdfplumber
joblib
requests
opencv-python
pymupdf
matplotlib